python_service/requirements.txt
tensorflow-cpu==2.12.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
opencv-python==4.8.1.78
Pillow==10.0.1
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai==0.5.4
python-dotenv==1.0.1